Master

+++ app/Http/Controllers/CertificateController.php
@@ -2,12 +2,16 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
+use App\Http\Requests\Certificate\CreateCertificateRequest;
 use App\Models\Certificate;
+use App\Models\Course;
+use App\Models\CourseMembershipDiscount;
 use App\Models\Subject;
 use Illuminate\Contracts\View\View;
-use App\Http\Requests\Certificate\CreateCertificateRequest;
-use App\Models\Course;
+use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class CertificateController extends Controller
 {
@@ -18,8 +22,8 @@ class CertificateController extends Controller
      */
     public function index()
     {
-        $certificates = Certificate::all();
-        return view('admin.certificate.index',[
+        $certificates = Certificate::orderBy('created_at', 'desc')->paginate(15);
+        return view('admin.certificate.index', [
             'certificates' => $certificates,
         ]);
     }
@@ -37,7 +41,7 @@ class CertificateController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  CreateCertificateRequest  $request
+     * @param CreateCertificateRequest $request
      * @return \Illuminate\Http\RedirectResponse
      */
     public function store(CreateCertificateRequest $request)
@@ -53,31 +57,84 @@ class CertificateController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  Certificate $certificate
+     * @param Certificate $certificate
      * @return View
      */
-    public function show(Certificate $certificate)
+    public function show(Certificate $certificate, Request $request)
     {
-        $subject = Subject::with('certificate')->get();
+        $input = $request->input();
+        $subjects = null;
+        if (empty($input['certificate'])) {
+            $subjects = Subject::with('certificate')->whereHas('certificate', function (Builder $query) {
+                return $query->where('id', 1);
+            })->get();
+        } else {
+            $subjects = Subject::with('certificate')->whereHas('certificate', function (Builder $query) use ($input) {
+                return $query->where('id', $input['certificate']);
+            })->get();
+        }
 
-        $courses = Course::with('subject','tutor', 'certificate')->get();
+        $_certificate = Certificate::with('subject')->where('id', $certificate->id)->first();
+        DB::enableQueryLog();
+        $courses_with_group = CourseMembershipDiscount::with(
+            'membershipCourses',
+            'courseDiscounts',
+            'membershipCourses.course',
+            'membershipCourses.course.subject',
+            'membershipCourses.course.subject.certificate',
+            'membershipCourses.course.tutor',
+            'membershipCourses.course.courseMaterial'
+        )->where('publish', 1)
+            ->whereHas('membershipCourses', function ($query) {
+                return $query->where('membership_id', Auth::check() ? Auth::user()->membership_group : 1);
+            })
+            ->whereHas('membershipCourses.course.subject.certificate', function ($query) use ($certificate) {
+                return $query->where('id', $certificate->id);
+            })->when(!empty($input['sort']), function ($query) use ($input) {
+                switch ($input['sort']) {
+                    case 'latest':
+                        return $query->orderBy('created_at', 'desc');
+                    case 'price':
+                        return $query->whereHas('membershipCourses.course', function ($query) {
+                            return $query->orderBy('course_price', 'asc');
+                        });
+                        break;
+                    case 'live':
+                        return $query->whereHas('membershipCourses.course', function ($query) {
+                            return $query->where('type', Course::$LIVE);
+                        });
+                        break;
+                    case 'record':
+                        return $query->whereHas('membershipCourses.course', function ($query) {
+                            return $query->where('type', Course::$RECORD);
+                        });
+                        break;
+                    case 'document':
+                        $query->whereHas('membershipCourses.course', function ($query){
+                            return $query->where('type', Course::$DOCUMENT);
+                        });
+                    default:
+                        break;
+                }
+            })->get();
+        // dd(DB::getQueryLog());
 
         return view('certificate.index', [
-            'certificate'=> $certificate,
-            'subject' => $subject,
-            'courses' => $courses,
+            'courses' => $courses_with_group,
+            'certificate' => $_certificate,
+            'subjects' => $subjects
         ]);
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  Certificate  $certificate
+     * @param Certificate $certificate
      * @return View
      */
     public function edit(Certificate $certificate)
     {
-        return view('admin.certificate.edit',[
+        return view('admin.certificate.edit', [
             'certificate' => $certificate
         ]);
     }
@@ -85,8 +142,8 @@ class CertificateController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  CreateCertificateRequest  $request
-     * @param  Certificate  $certificate
+     * @param CreateCertificateRequest $request
+     * @param Certificate $certificate
      * @return \Illuminate\Http\RedirectResponse
      *
      */
@@ -102,11 +159,49 @@ class CertificateController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  Certificate  $certificate
+     * @param Certificate $certificate
      * @return \Illuminate\Http\Response
      */
     public function destroy(Certificate $certificate)
     {
-        $certificate->delete();
+//        try {
+//            $_subjects = $certificate->subject->toArray();
+//            if (empty($_subjects)){
+//                $certificate->delete();
+//            }
+//            return response([
+//                'message' => 'Delete success!'
+//            ],200);
+//        } catch (\Exception $exception) {
+//            return response([
+//                'message' => 'Cannot delete',
+//                'error' => $exception->getMessage()
+//            ], 400);
+//        }
+         DB::beginTransaction();
+         try {
+//             $_courses = Course::whereHas('subject.certificate', function ($query) use ($certificate){
+//                 return $query->where('id', $certificate->id);
+//             });
+//             $_subjects = Subject::whereHas('certificate', function ($query) use ($certificate){
+//                 return $query->where('id', $certificate->id);
+//             })->get();
+//             $_courses->delete();
+//             $_subjects->delete();
+             $_subjects = $certificate->subject->toArray();
+             if (empty($_subjects)){
+                 $certificate->delete();
+                 DB::commit();
+                 return response([
+                     'message' => 'Delete success!'
+                 ]);
+             }
+         } catch (\Throwable $th) {
+             DB::rollback();
+             return response([
+                 'message' => 'Cannot delete',
+                 'detail' => $th->getMessage()
+             ], 400);
+         }
     }
 }
